chore: rename utils.file to files
